### 2024-02-19 #6376 @editorialbot
[PRE REVIEW]: eVAS: A user-friendly electronic Visual Analogue Scale
**Submitting author:** @gouverneurp (<a href="http://orcid.org/0000-0001-8610-2291">Philip Gouverneur</a>)
**Repository:** https://github.com/gouverneurp/eVAS/
**Branch with paper.md** (empty if default branch):
**Version:** 0.1.0
