### 2024-02-20 d348c4b
Added DOIs to paper references
