+++ paper/paper.bib
@@ -6,6 +6,7 @@
 	number    = {3},
 	pages     = {394--404},
 	year      = {2022},
+	doi       = {10.1007/s43441-022-00376-2},
 	publisher = {Springer}
 }
 
@@ -17,6 +18,7 @@
 	number    = {4},
 	pages     = {558--560},
 	year      = {2006},
+	doi       = {10.1038/sj.ejcn.1602342},
 	publisher = {Nature Publishing Group}
 }
 
@@ -28,6 +30,7 @@
 	number    = {9},
 	pages     = {1215--1225},
 	year      = {2017},
+	doi       = {10.1177/0269215517692641},
 	publisher = {SAGE Publications Sage UK: London, England}
 }
 
@@ -38,6 +41,7 @@
 	volume    = {2},
 	number    = {3},
 	year      = {2018},
+	doi       = {10.5435/jaaosglobal-d-17-00088},
 	publisher = {Wolters Kluwer Health}
 }
 
@@ -49,6 +53,7 @@
 	number    = {3},
 	pages     = {565--574},
 	year      = {2021},
+	doi       = {10.1007/s11245-020-09724-z},
 	publisher = {Springer}
 }
 
@@ -60,6 +65,7 @@
 	number    = {1},
 	pages     = {67--73},
 	year      = {2009},
+	doi       = {10.1111/j.1471-0528.2008.02002.x},
 	publisher = {Wiley Online Library}
 }
 
@@ -71,6 +77,8 @@
 	number    = {4},
 	pages     = {1045--1047},
 	year      = {2008},
+	doi       = {10.1016/j.pnpbp.2008.02.002},
+
 	publisher = {Elsevier}
 }
 
@@ -82,6 +90,7 @@
 	number    = {1},
 	pages     = {99--106},
 	year      = {2009},
+	doi       = {10.3758/BRM.41.1.99},
 	publisher = {Springer}
 }
 
@@ -91,6 +100,7 @@
 	journal   = {Information Technology in Biomedicine},
 	pages     = {239--251},
 	year      = {2021},
+	doi       = {10.1007/978-3-030-49666-1_19},
 	publisher = {Springer}
 }
 
@@ -102,6 +112,7 @@
 	number    = {14},
 	pages     = {4838},
 	year      = {2021},
+	doi       = {10.3390/s21144838},
 	publisher = {MDPI}
 }
 
@@ -121,6 +132,7 @@
 	journal   = {Quality of Life Research},
 	pages     = {1--11},
 	year      = {2023},
+	doi       = {10.1007/s11136-023-03411-3},
 	publisher = {Springer}
 }
 
@@ -132,6 +144,7 @@
 	number    = {1},
 	pages     = {67--75},
 	year      = {2016},
+	doi       = {10.1016/j.sjpain.2016.06.012},
 	publisher = {De Gruyter}
 }
 
@@ -143,6 +156,7 @@
 	number    = {9},
 	pages     = {e18284},
 	year      = {2020},
+	doi       = {10.2196/18284},
 	publisher = {JMIR Publications Toronto, Canada}
 }
 
@@ -154,6 +168,7 @@
 	number    = {4},
 	pages     = {447--453},
 	year      = {2014},
+	doi       = {10.1016/j.jbtep.2014.06.004},
 	publisher = {Elsevier}
 }
 
@@ -165,6 +180,7 @@
 	number    = {1},
 	pages     = {228--237},
 	year      = {2024},
+	doi       = {10.1016/j.jpain.2023.08.005},
 	publisher = {Elsevier}
 }
 
@@ -176,6 +192,7 @@
 	number    = {1},
 	pages     = {e0280579},
 	year      = {2023},
[…]
