### 2024-02-20 #6376 @gouverneurp
From the list of similar articles, [PsychoPhysioCollector: A Smartphone-Based Data Collection App for Psychophysiological Research](https://github.com/openjournals/joss-reviews/issues/40) seems to be the best match in terms of content. So @sbogutzky ([Simon Bogutzky](http://orcid.org/0000-0002-9279-6549)) and @davclark might be a good fit for reviewing **eVAS**. Unfortunately, the programming languages of the two projects are different.
