### 2024-02-20 #6376 @gouverneurp
From [the list of people](https://reviewers.joss.theoj.org/lookup), several people could be candidates for reviewers.
For example:
[vankesteren](https://github.com/vankesteren)
[bbortey9](https://github.com/bbortey9)
[sneakers-the-rat](https://github.com/sneakers-the-rat)
[JerryChiaRuiChang](https://github.com/JerryChiaRuiChang)
