### 2024-02-21 3ce3dd6
Fixed path to scripts in README
