+++ README.md
@@ -3,7 +3,7 @@
 [![status](https://joss.theoj.org/papers/fcfce9c9fb4d33c87dac8cf876fd3b27/status.svg)](https://joss.theoj.org/papers/fcfce9c9fb4d33c87dac8cf876fd3b27)
 [![License: MIT](https://img.shields.io/badge/License-MIT-yellow.svg)](https://github.com/gouverneurp/eVAS/blob/main/LICENSE.MD)
 
-# eVAS
+# eVAS: A user-friendly electronic Visual Analogue Scale
 <picture>
     <img src="./images/icon.png?raw=true" width="50"/>
 </picture>
