### 2024-02-23 #6376 @davclark
I don't tend to be that active on GitHub these days, so I apologize in advance that I'm likely to be slow in responding. But I agree that I have a strong understanding of this space. I want to endorse the notability of the offering given the fact that VAS is basically psychometrically superior to the ubiquitous likert scale, but everyone still uses likert scales. I can probably provide refs if you really want, but as an ex-academic, I of course have limited access to the literature! That said, the paper provides at least one reference on the value of the VAS.

I actually wrote my own multi-lingual system for displaying forms, including VAS. It would take definitions like [this](https://github.com/glass-bead-labs/simple-forms/blob/master/form-data/ai-a-nz.yaml) and render them with [this logic](https://github.com/glass-bead-labs/simple-forms/blob/master/src/simple_forms/core.clj#L29-L42). It was a pretty bespoke / custom setup, including the fact that I hand-edited [the css](https://github.com/glass-bead-labs/simple-forms/blob/master/rendered/simple-forms-style.css) that was located in the otherwise computer-generated outputs. Tl;dr, implementing the VAS is non-trivial as it is unfortunately not a common metric - even qualtrics doesn't provide the level of easy customization of eVAS.

eVAS provides a rich framework for displaying various kinds of VAS questions. The only thing I didn't see was the ability to hide the thumb (slider widget) prior to a press, as this can anchor responses. I had a quick read of the paper, and I salute the authors for packaging their software to be simple and accessible to non-programmers.

I unfortunately do not have time to evaluate the software directly. If someone else could help with this, that would be ideal. But if needed, I can try to find time perhaps next week or so.
