### 2024-02-26 #6376 @gouverneurp
Thank you for your reply and feedback!
Would you be so kind as to comment in more detail on what you meant by "The only thing I didn't see was the ability to hide the thumb (slider widget) prior to a press, as this can anchor responses". Once I fully understand the idea, I can implement it and add it to eVAS.
