### 2024-02-27 #6376 @davclark
I had a quick look and couldn't find a citation. I'm guessing you know the general idea of anchoring, but basically, any identifiable location in a response field will tend to "magnetize" the distribution of responses towards that location.

So, one idea is that you hide the slider (called a thumb in browser land) until someone clicks or touches the range slider. CSS for webkit-compatible browsers is here:

https://github.com/glass-bead-labs/simple-forms/blob/master/rendered/simple-forms-style.css#L39-L47

Docs here: https://developer.mozilla.org/en-US/docs/Web/CSS/::-webkit-slider-thumb

Certainly, I would not consider that feature a blocker for publication! It's very nice how you've provided a great deal of customization.
