### 2024-02-28 25b744d
Implement "on_click" & "hide_slider" options and pnly use mouse updates dependent on "step_size"
