+++ eVAS.py
@@ -162,6 +162,8 @@ def create_config():
     config.set("devices", "move_while_down", "False")
     config.set("devices", "# Whether to use the mouse (slider does not move with keys anymore). Should be: True/False")
     config.set("devices", "use_mouse", "False")
+    config.set("devices", "# Whether to update only on mouse click (only has an effect if 'use_mouse' is 'True'.). Should be: True/False")
+    config.set("devices", "on_click", "False")
 
     config.add_section("keys")
     config.set("keys", "# The following section is used to defined the keys.")
@@ -199,6 +201,8 @@ def create_config():
     config.set("appearance", "use_two_triangle", "False")
     config.set("appearance", "# Welcome message that is displayed at the beginning. Usually, you do not want to change this.")
     config.set("appearance", "welcome_message", f"Press space to start. You can configure '{config_file_name}' to change the eVAS.")
+    config.set("appearance", "# Whether to hide the slider until first user interaction. Should only be used with 'use_mouse' set to 'True'. Should be: True/False.")
+    config.set("appearance", "hide_slider", "False")
 
     config.add_section("csv")
     config.set("csv", "# Delimiter used in the CSV output file, for example ';'.")
@@ -302,7 +306,11 @@ def eval_config(config):
 # CoVAS
 class Covas(tk.Frame):
     def __init__(self, callback, master=None, *args, **kwargs):
-        super().__init__(master, cursor='none', background="white")
+
+        on_click = eval(config["devices"]["on_click"])
+        # set a visible cursor if interact by mouse click, otherwise set cursor invisible
+        cursor = "target" if on_click else "none"
+        super().__init__(master, cursor= cursor, background="white")
         self.pack(fill=BOTH, expand=True)
         self.focus_set()
         self.callback = callback
@@ -340,6 +348,7 @@ class Slider(tk.Canvas):
             self.use_mouse = eval(config["devices"]["use_mouse"])
             self.move_while_down = eval(config["devices"]["move_while_down"])
             self.trigger_thermode = eval(config["devices"]["trigger_thermode"])
+            self.on_click = eval(config["devices"]["on_click"])
 
             self.keys_start = eval(config["keys"]["keys_start"])
             self.keys_end = eval(config["keys"]["keys_end"])
@@ -357,6 +366,7 @@ class Slider(tk.Canvas):
             self.slider_height = eval(config["appearance"]["slider_height"])
             self.slider_color = eval(config["appearance"]["slider_color"])
             self.welcome_message = config["appearance"]["welcome_message"]
+            self.hide_slider = eval(config["appearance"]["hide_slider"])
         except Exception as e:
             throw_config_error(e)
 
@@ -370,7 +380,10 @@ class Slider(tk.Canvas):
         self.bind('<Configure>', self.update_size)
 
         if self.use_mouse:
-            self.bind('<Motion>', self.update_mouse)
+            if self.on_click:
+                self.bind('<ButtonRelease-1>', self.update_mouse)
+            else:
+                self.bind('<Motion>', self.update_mouse)
 
         if self.trigger_thermode:
             self.com = get_com()
@@ -429,12 +442,12 @@ class Slider(tk.Canvas):
 
         # visualize the slider background
         self.delete('all')
-        gradient_w, gradient_h = int(w*(1-2*xpad)+.5), int(h*(1-2*ypad)*gradh)
+        self.gradient_w, gradient_h = int(w*(1-2*xpad)+.5), int(h*(1-2*ypad)*gradh)
         gradient_y = h*(ypad+(1-2*ypad))*gradfac
         if self.use_image:
             image_path = self.get_image_path()
[…]
