+++ examples/simplified_faces_pain_scale/config.ini
@@ -3,8 +3,6 @@
 sampling_rate = 0.1
 # Whether to save values only upon change. Should be: True/False
 only_on_change = True
-# Whether to check for updates. Should be: True/False
-check_for_updates = True
 # Whether to use the second screen to show the eVAS. Should be: True/False
 use_second_screen = True
 
@@ -17,6 +15,8 @@ label_size = 32
 range = [0, 2]
 # Define the start value
 start = 0
+# Step size to move the slider whenever a key is released.
+step_size = 1
 # Whether to plot vertical lines for each anchor. Should be: True/False
 vertical = True
 # Whether to plot numbers below the lines. Should be: True/False
@@ -25,12 +25,12 @@ numbers = False
 [devices]
 # Whether to trigger the thermode (QST.LAB TCS2). Should be: True/False
 trigger_thermode = False
-# Step size to move the slider whenever a key is released.
-step_size = 1
 # Moving the slider not only when the button is released, but also while the button is held down. Should be: True/False
 move_while_down = False
 # Whether to use the mouse (slider does not move with keys anymore). Should be: True/False
 use_mouse = False
+# Whether to update only on mouse click (only has an effect if 'use_mouse' is 'True'.). Should be: True/False
+on_click = False
 
 [keys]
 # The following section is used to defined the keys.
@@ -65,5 +65,15 @@ mid_color = None
 use_triangle = False
 # Use a decreasing and an increasing triangle instead of a square for the background. Should be: True/False.
 use_two_triangle = False
+# Welcome message that is displayed at the beginning. Usually, you do not want to change this.
 welcome_message = Press space to start. You can configure 'config.ini' to change the eVAS.
+# Whether to hide the slider until first user interaction. Should only be used with 'use_mouse' set to 'True'. Should be: True/False.
+hide_slider = False
 
+[csv]
+# Delimiter used in the CSV output file, for example ';'.
+delimiter = ;
+# Decimal point used in the CSV output file, for example ','.
+decimal_point = ,
+# Number of digits after the decimal point for the slider values saved.
+decimal_places = 4
