### 2024-02-28 1699c7c
Added new parameters to manuscript
