### 2024-03-10 #6376 @Kevin-Mattheus-Moerman
@gouverneurp I have just removed the `query-scope` label as this work was found to potentially be in scope for JOSS. We will now aim to find and assign a handling editor.
