### 2024-03-14 #6376 @britta-wstnr
Hi @davclark - it looks like you already had a good look at this software and you already gave some feedback. We are in not hurried right now to start the review process, given that we are still looking for a sufficient number of reviewers to even be able to start that step. Do you think you would be able to formally review this for us some time in the upcoming weeks?

Here at JOSS we use a check-box system for reviews that would help you along in the process.
