### 2024-04-25 27d07f9
Set execution permission & stop on error
