### 2024-04-25 22e3a84
Wait for other software managers to finish in ubuntu scripts
