+++ scripts/create_ubuntu_app.sh
@@ -1,5 +1,12 @@
 #!/bin/bash
 
+# stop on error
+set -e
+
+############################################################
+# Functions                                                #
+############################################################
+
 function wait_until_free() {
     i=0
     tput sc
@@ -25,8 +32,9 @@ function install() {
     echo "____________________________________"
 }
 
-# stop on error
-set -e
+############################################################
+# Main                                                     #
+############################################################
 
 cd "$(dirname "$0")"
 
