### 2024-06-26 gouverneurp/eVAS#3 @haoxue-fan
Conflict between installing eVAS via python and standalone (Mac)
I first installed eVAS through the standalone app for Mac, which worked great. I then tried to install it through the pip install method, and when I ran `python eVAS.py`, I saw the message 'You need to add eVAS to the Accessibility and Input Monitoring apps and restart it. Open the system preferences?' even though I have already added the eVAS standalone to the allowed list. I wonder whether there is anyway to prevent this warning, or maybe just make it clearer in the README that people should only use one of the way to install eVAS?
