### 2024-06-26 gouverneurp/eVAS#6 @haoxue-fan
Unexpected exiting when pressing 'caps lock' on Mac standalone app
The Mac standalone app exits after I press caps lock. There will be one data file saved, but no data inside (i.e., the size for the file is 0 byte)

I am using Mac Sonoma 14.2.1 and python 3.11.4.
