### 2024-06-26 5900889
Possibility to save the config at the beginning of the output CSV
