### 2024-06-27 552baee
Added the possibility to change font size of numbers
