### 2024-06-27 937447e
Round values and remove tailing zeros
