### 2024-06-27 49d398c
Fix for "send_trigger" bug
