### 2024-06-27 9c44df0
Updated README about building executables
