### 2024-07-18 #6876 @amitchell12
I agree that open source code for an easy to use VAS to rate pain is extremely useful, as well as a way of standardising VASs within pain research. It is excellent that it is available for download on all platforms, but this should be made clear at the top of the ReadME (I was initially confused as thought it might just run on windows). I appreciate the ease of use and the fact that the results are immediately saved in a .csv file. I have some comments with regard to functionality and performance:

The eVAS APP downloaded onto the mac without issue, and worked as it should, but I could not easily exit the VAS with a button press, and only found out it was "ESc" by reading the associated paper. This information should be in the README and in the start-up text when you run the application.

It is not clear from the README that the VAS is continuous - only from the output, this should be clear from the instructions that the response is continuously recording.

The output .csv needs perfecting. The title of the columns are not seperated. The third column - which has values in it - is not named.

The statement of need should be present in application documentation, not just the paper.

Examples are provided but it is not clear from the README how the default VAS can be updated to make use of the different examples (I find the cold warm one particularly useful). Documentation should be clearer on how to set these VAS alternatives up.

I do not understand why the default scale is in a colour gradient. It is arbritary and does not really protray (in my opinion) the desired increase in pain intensity that I assume the authors are going for. For me, it is just a distraction. It is much more appropriate for the VAS to visualised be a single, thin black line with the weights of 'not at all' and 'worst pain imaginable' on the left and right respectively. The colour coding for the warm and cold VAS makes sense.

Finally, whilst I agree that standardisation is required. This eVAS software does not address some of the more pressing issues with respect to VAS ratings - namely that standardisation issues actually depend more on instruction and individual interpretation of the scale. If the authors wish for more standardisation, they should include a clear section of guidelines for users on how to instruct participants/patients to use the VAS - this is missing and I believe it would be an incredibly useful addition.
