### 2024-07-26 #6876 @haoxue-fan
Second to the review from @amitchell12 RE a more clear readme structure and more documentation on how to customize the appearance/scale. It is indeed helpful to have a user friendly eVAS given its prevelance in social science research.

I have raised several issues in the original repo regarding adding features and adjusting the existing interface, which the author has mostly perfectly resolved. There are just a couple more issues left open in the repo regarding the usage of the standalone version of Mac. Totally understand that it may take longer time to resolve since the author has mentioned that they need to compile everything to make an updated version of the standalone app, and I think it is also fine with me if the author add a couple of lines in README acknowledging these issues (since they are not fatal) so that the users are aware of them. Overall, great job!
