### 2024-08-11 44f53ac
Ensure that escape is in list of keys to end eVAS
