### 2024-08-11 371a1b5
Added more info about how to use and added info about how to end the eVAS on the welcome message
