### 2024-08-11 64c080a
Added statement of need in README
