### 2024-08-11 6554bdd
Added link for cold hot scale
