+++ examples/custom_image_scale/config.ini
@@ -21,15 +21,17 @@ step_size = 0.02
 vertical = False
 # Whether to plot numbers below the lines. Should be: True/False
 numbers = False
+# Font size of the defined numbers below the scale.
+number_size = 32
 
 [devices]
-# Whether to trigger the thermode (QST.LAB TCS2). Should be: True/False
+# Whether to send a signal to trigger a 'QST.LAB TCS2' thermode when the eVAS is recording. Only works on Windows. Should be: True/False
 trigger_thermode = False
 # Moving the slider not only when the button is released, but also while the button is held down. Should be: True/False
 move_while_down = False
 # Whether to use the mouse (slider does not move with keys anymore). Should be: True/False
 use_mouse = False
-# Whether to update only on mouse click (only has an effect if 'use_mouse' is 'True'.). Should be: True/False
+# Whether to update only on mouse click (only has an effect if 'use_mouse' is 'True'). Should be: True/False
 on_click = False
 
 [keys]
@@ -65,8 +67,8 @@ mid_color = None
 use_triangle = False
 # Use a decreasing and an increasing triangle instead of a square for the background. Should be: True/False.
 use_two_triangle = False
-# Welcome message that is displayed at the beginning. Usually, you do not want to change this.
-welcome_message = Press space to start. You can configure 'config.ini' to change the eVAS.
+# Welcome message that is displayed at the beginning. Should be list of Strings. Each element of the list is separated by a line break.
+welcome_message = [f"Press {get_key_string(self.keys_start)} to start.", f"You can configure {config_file_name} to change the eVAS.", f"The application can be exited at any time by pressing {get_key_string(self.keys_end)}."]
 # Whether to hide the slider until first user interaction. Should only be used with 'use_mouse' set to 'True'. Should be: True/False.
 hide_slider = False
 
@@ -77,4 +79,6 @@ delimiter = ;
 decimal_point = ,
 # Number of digits after the decimal point for the slider values saved.
 decimal_places = 4
+# Whether to save the used config at the beginning of the CSV file.
+save_config_in_csv = False
 
