### 2024-08-11 5233dc3
Moved screenshot images of the examples
