### 2024-08-11 68f2082
Added config for simple line
