### 2024-08-11 563cb2b
Customize vertical lines and height of background gradient
