+++ examples/simplified_faces_pain_scale/config.ini
@@ -17,8 +17,8 @@ range = [0, 2]
 start = 0
 # Step size to move the slider whenever a key is released.
 step_size = 1
-# Whether to plot vertical lines for each anchor. Should be: True/False
-vertical = True
+# Height of vertical lines for each anchor. Can be 0 to not draw any vertical lines.
+vertical_line_height = 50
 # Whether to plot numbers below the lines. Should be: True/False
 numbers = False
 # Font size of the defined numbers below the scale.
@@ -63,6 +63,8 @@ left_color = (256,256,256)
 right_color = (256,256,256)
 # Background color mid. Can also be disabled with 'None'.
 mid_color = None
+# Height of the line the slider moves on.
+line_height = 128
 # Use a triangle instead of a square for the background. Should be: True/False.
 use_triangle = False
 # Use a decreasing and an increasing triangle instead of a square for the background. Should be: True/False.
