+++ README.md
@@ -130,6 +130,10 @@ All help is welcome and needed! Feel free to open pull requests and contact us v
 # Report issues or problems with the software
 You are welcome to open issues here directly on GitHub, or contact us by email at [philipgouverneur@gmx.de](mailto:philipgouverneur@gmx.de).
 
+# Known Issues
+
+- Pressing the Caps Lock key may cause the application to crash on certain versions of MacOS. The problem is known for eVAS and is strongly related to tkinter not working properly in combination with pynput.
+
 # Used resources
 The application icon (<picture><img src="./images/icon.png?raw=true" width="12"/></picture>) is free to use and can be found at the following [link](https://de.freepik.com/icon/schlecht_10012613#fromView=search&term=pain+rating&page=1&position=16&track=ais).
 
