### 2024-08-11 gouverneurp/eVAS#3 @gouverneurp
To be honest, I cannot really reproduce this problem. Is it related to the fact that you also need to add Python to the Accessibility and Input Monitoring applications?
What would you recommend I do?
