### 2024-08-11 gouverneurp/eVAS#6 @gouverneurp
I can actually reproduce this bug. I think it is closely related to problems between Tkinter and pynput (bibliography to detect keystrokes). I have not found much online to solve the problems. I will commit a change that will reduce possible crashes, but not fix them completely in the near future.
