### 2024-08-11 477d30c
Reduce crashes on mac
