### 2024-08-11 gouverneurp/eVAS#6 @gouverneurp
As mentioned, I have been working on a fix for the last days.
Here is the possible solution:
https://github.com/gouverneurp/eVAS/commit/477d30c255570a640163faef09c941c755c02104

Can you check it again?
I will also release a built version later on.
