### 2024-08-11 #6876 @gouverneurp
Thanks again for your valuable help and feedback!
Following your comments, I have released a new version: [0.2.0](https://github.com/gouverneurp/eVAS/releases/tag/0.2.0)!

In the following, I will try to address and comment on your feedback individually:
- Shutdown information: You can now exit the application using the Escape key, regardless of the configuration you have chosen. In addition, the buttons to exit the application are mentioned at the start of the application and in the README of the github page. [commit](https://github.com/gouverneurp/eVAS/commit/371a1b553f27031c4035854f99695855b79b6485)
- Continuing: I have added more information to the README. I hope this makes it clearer that the eVAS is continuous. [commit](https://github.com/gouverneurp/eVAS/commit/371a1b553f27031c4035854f99695855b79b6485)
- CSV: I checked the output CSV again and it works fine for me and displays a valid CSV file.
Example Textfile:
![grafik](https://github.com/user-attachments/assets/30a15c81-06cc-495d-a5e3-a7fcfd9ef35e)
Resulting CSV:
![grafik](https://github.com/user-attachments/assets/225b480c-9134-4349-beb7-f15a6be2484f)
If you still have problems, please check the CSV configuration. The default configuration is set up for European CSV files (delimiter is a semicolon, decimal point is a comma), but you can change it as needed (American with delimiter as comma and decimal point as a single point?):
```
[csv]
# Delimiter used in the CSV output file, for example ';'.
delimiter = ;
# Decimal point used in the CSV output file, for example ','.
decimal_point = ,
# Number of digits after the decimal point for the slider values saved.
decimal_places = 4
# Whether to save the used config at the beginning of the CSV file.
save_config_in_csv = False
```

- Statement of Need: The Statement of Need has been added to the README file. [commit](https://github.com/gouverneurp/eVAS/commit/64c080a3d1501963dd453ce96f9f966c7e5ecfcf)
- Examples: It is now easier to implement the examples. You can click on the images of the examples and find the folder containing the configuration directly. The next step is to copy the files. A detailed explanation has been added to the README file. [commit](https://github.com/gouverneurp/eVAS/commit/13b27106f12e4c35164b02d4f3dae290a82fa4e0)
- We believe that one of the strengths of **eVAS** is its ability to display any scale you require. So we have added another example configuration for a simple black line as explained. [commit](https://github.com/gouverneurp/eVAS/commit/68f2082824d04d9e65d67aeef53bc8f7b1d42fed)
- Standardisation: We believe that **eVAS** can help standardise the collection of continuous linear measurements by providing a simple tool to collect measurements in a standardised way. Any standardisation of a study protocol to collect data from subjects would be beyond the scope of this application. How can we add it here?
- Github issues: All open issues in the github repository have been answered with either a possible solution with a commit or a comment. [issue6](https://github.com/gouverneurp/eVAS/issues/6) [issue3](https://github.com/gouverneurp/eVAS/issues/3)

Please let us know if you have any other ideas for improving **eVAS**!
