+++ eVAS.py
@@ -516,14 +516,15 @@ class Slider(tk.Canvas):
 
         # visualize the labels with optional bars and numbers
         y2 = (1-textfac)*gradient_y+semi_line_height + textfac*h*(1-ypad)
+        offset = 20
         for i, x, text in zip(np.linspace(start= self.range[0], stop= self.range[1], num=len(self.labels)), range(int(xs[0]), int(xs[-1]), int((xs[-1] -  xs[0]) // max((len(self.labels) - 1), 1)) - 1), self.labels):
-            self.create_text(x, y2, text=text, font=('DejaVu',self.label_size), anchor=CENTER, justify='center', fill='black')
+            self.create_text(x, y2 + self.vertical_line_height + offset, text=text, font=('DejaVu',self.label_size), anchor=CENTER, justify='center', fill='black')
             # vertical lines
             y0,y1 = gradient_y-self.vertical_line_height, gradient_y+self.vertical_line_height
             self.create_line(*[x,y0, x,y1], fill='#000000', width=3)
             if self.numbers:
                 value = str(round(i, 2)).rstrip('0').rstrip('.')
-                self.create_text(x, y1, text= value, font=('DejaVu', self.number_size, 'bold'), anchor=N, fill='black')
+                self.create_text(x, y1 + offset, text= value, font=('DejaVu', self.number_size, 'bold'), anchor=N, fill='black')
 
         # create slider already if it should not be invisible at the start
         if not self.hide_slider:
