### 2024-08-12 df9e386
Added numeric rating scale (NRS) example
