+++ examples/numerical_rating_scale/config.ini
@@ -1,0 +1,86 @@
+[general]
+# Sampling rate: How often should values be saved to file (in seconds).
+sampling_rate = 0.1
+# Whether to save values only upon change. Should be: True/False
+only_on_change = True
+# Whether to use the second screen to show the eVAS. Should be: True/False
+use_second_screen = True
+
+[scale]
+# Define the anchor labels of the scale. You can wrap text by placing a '\n' in the label.
+anchors = ['Not at all', '', '', '', '', '', '', '', '', '', 'Worst\nimaginable pain']
+# Font size of the defined anchor labels above.
+label_size = 32
+# Define the value range
+range = [0, 10]
+# Define the start value
+start = 0
+# Step size to move the slider whenever a key is released.
+step_size = 1
+# Height of vertical lines for each anchor. Can be 0 to not draw any vertical lines.
+vertical_line_height = 50
+# Whether to plot numbers below the lines. Should be: True/False
+numbers = True
+# Font size of the defined numbers below the scale.
+number_size = 32
+
+[devices]
+# Whether to send a signal to trigger a 'QST.LAB TCS2' thermode when the eVAS is recording. Only works on Windows. Should be: True/False
+trigger_thermode = False
+# Moving the slider not only when the button is released, but also while the button is held down. Should be: True/False
+move_while_down = False
+# Whether to use the mouse (slider does not move with keys anymore). Should be: True/False
+use_mouse = False
+# Whether to update only on mouse click (only has an effect if 'use_mouse' is 'True'). Should be: True/False
+on_click = False
+
+[keys]
+# The following section is used to defined the keys.
+# Keys code names can be found under: 'https://pynput.readthedocs.io/en/latest/keyboard.html'
+# Keys to start the recording.
+keys_start = [Key.space]
+# Keys to end the application.
+keys_end = [Key.esc, Key.end]
+# Keys to move the slider left.
+keys_left = [Key.left, Key.page_up]
+# Keys to move the slider right.
+keys_right = [Key.right, Key.page_down]
+
+[appearance]
+# Whether to load an image as slider background. Image must be placed in the same directory and called 'image'. Should be: True/False
+use_image = False
+# Whether to load an image and place it above the slider. Image must be placed in the same directory and called 'upper_image'. Should be: True/False
+use_upper_image = False
+# Width of the slider
+slider_width = 31
+# Height of the slider
+slider_height = 138
+# Color of the slider. Should be in form (R,G,B). Colors can be found under: 'https://redketchup.io/color-picker'
+slider_color = (255,0,0)
+# Background color left
+left_color = (0,0,0)
+# Background color right
+right_color = (0,0,0)
+# Background color mid. Can also be disabled with 'None'.
+mid_color = None
+# Height of the line the slider moves on.
+line_height = 4
+# Use a triangle instead of a square for the background. Should be: True/False.
+use_triangle = False
+# Use a decreasing and an increasing triangle instead of a square for the background. Should be: True/False.
+use_two_triangle = False
+# Welcome message that is displayed at the beginning. Should be list of Strings. Each element of the list is separated by a line break.
+welcome_message = [f"Press {get_key_string(self.keys_start)} to start.", f"You can configure {config_file_name} to change the eVAS.", f"The application can be exited at any time by pressing {get_key_string(self.keys_end)}."]
+# Whether to hide the slider until first user interaction. Should only be used with 'use_mouse' set to 'True'. Should be: True/False.
+hide_slider = False
+
+[csv]
+# Delimiter used in the CSV output file, for example ';'.
+delimiter = ;
+# Decimal point used in the CSV output file, for example ','.
+decimal_point = ,
[…]
