### 2024-08-19 gouverneurp/eVAS#3 @haoxue-fan
I am not sure what the root issue is, but on a second thought I guess it may not be too common for people to both install it as standalone and through command line (or if they do that, they may be able to figure out how to make it work by probably de-installing the standalone version). I think this was just a minor point (and I happen to encounter this during testing), so I would recommend just to close this issue for now!
