### 2024-12-09 #6876 @britta-wstnr
Hi @gouverneurp - find below my points,  evaluating the implementation of @amitchell12’s reviewer points since we could not reach this reviewer again.
Most points are sufficiently implemented from my view (✅). Some could use some work (✏️). And then I ran into a problem installing eVAS, so two points I could not evaluate yet (⛔). More details on the installation problem below.

### Revision: implementation
1. Document prominently on which platforms eVAS runs
 This could still be improved by adding a sentence to the first paragraph of the README, listing the platforms eVAS runs on and linking to the installation explanation further down in the text. ✏️
2. How to exit eVAS on a Mac
This has been added to the documentation ✅
3. eVAS being continuous
This could still be made a little clearer in the README: for example, is eVAS _always_ continuous, even if there is a numerical scale with anchors 1-10 (bottom right example)? ✏️
4. csv file headers
not verified yet - see below “Problems executing eVAS” ⛔
5. statement of need
 Is added to the README now. It might benefit from a few references. ✅ + ✏️
6. How to set up eVAS configuration
not verified yet - see below “Problems executing eVAS” ⛔
7. Colour gradient default scale
 More scales were added, also black-white options. ✅
8. Standardisation
 @gouverneurp writes: “We believe that eVAS can help standardise the collection of continuous linear measurements by providing a simple tool to collect measurements in a standardised way. Any standardisation of a study protocol to collect data from subjects would be beyond the scope of this application. How can we add it here?”
  Maybe it could be an option to add a “Notes” section? There you could caution that the output of eVAS is also depending on a standardized instruction of the participants. ✏️

### Problems executing eVAS
In the process, I was trying to run eVAS (again). I remember not having any problems on the same machine (MacOS), when I first tried this before inviting reviewers. Now, I cannot get past running the `.command` file (I certainly updated my machine in between, but cannot tell you what version of MacOS I was running back then.)
-  I could not double-click the file (as you describe) since MacOS would simply refuse this (giving me an error without options), I could only get to the dialogue you show via right-click —> “Open”
- trying to then run the `.command` file,  I had to make the file an executable via the terminal first.
- Once that is done, I still cannot open eVAS, my Mac says simply “The application “eVAS.app” can’t be opened.” with no options. Adapting rights did not help for me.

I am currently on Sonoma, using an Apple M1 chip and MacBook Air.

### Some further comments one the README from me (only minor things):
- “easy to use and user-friendly slider”: add the hyphen between user and friendly
- “fear (of childbirth)” - that seems very specific, is this a necessary/useful addition here?
- “More specifically, papers using” - papers should be plural
