### 2024-12-28 #6876 @gouverneurp
Hi @britta-wstnr,
Thanks again for your valuable feedback!
We have tried to address all of your comments:
- eVAS on a Mac : Indeed, there seems to have been a problem with the Mac standalone version in the last release. As mentioned before, it was difficult to build the application for Mac because we had limited access to the operating system and were using virtual machines. To avoid problems here, we added a [git action](https://github.com/gouverneurp/eVAS/blob/main/.github/workflows/build_macos.yml) to build the application natively on MacOs using git. I have updated the application on the release page. Would it be possible for you to test it again?
- Document prominently on which platforms eVAS runs: We added the information in the README ([commit](https://github.com/gouverneurp/eVAS/commit/d0fd78687b7370fd5e193599862b13d2fef703ea))
- eVAS being continuous : More information was given here as well ([commit](https://github.com/gouverneurp/eVAS/commit/cb678231113214c83fa26d3c8594d871a695845f))
- Standardisation: A new section with a short note was added ([commit](https://github.com/gouverneurp/eVAS/commit/b76be53ca7167fe11d809e11d5524e7eae031847))
- Minor comments about README: We tried to fix the smaller issues ([commit](https://github.com/gouverneurp/eVAS/commit/fb8e7f99045164ad608d621332df2c8f469cee1e))
- Section Known issues: We added more information in the section ([commit](https://github.com/gouverneurp/eVAS/commit/3e34bbdc2693a51a95e5ae42cd98f16feafbc57f))
- section Used resources: can you report the license of the artwork you are using?
Unfortunately, there is no license to link. We added more informationen that the resource is from freepik ([commit](https://github.com/gouverneurp/eVAS/commit/feeaaa573d458c1e349f54916b4169f60ab41a02)). Doing so we follow the [rules](https://support.freepik.com/s/article/Attribution-How-when-and-where?language=en_US) of [freepik](https://freepik.com).

Please let us know if there is anything else we can do to improve **eVAS**!
