### 2025-01-13 #6876 @britta-wstnr
Hi @gouverneurp 👋
I tried to follow the MacOS instructions, however, on Sonoma, I cannot find the option for `run.command`. At the end of this week, I can update to Sequoia and try again and let you know (until then I have to keep this OS to not jeopardize some of my work). Which OS version did you work with?

I'll get to your other points together with the feedback on the install then 🙂
