### 2025-01-15 #6876 @gouverneurp
Hi @britta-wstnr,
We tested eVAS on _macOS BigSur_ using a virtual machine and on _macOS Sonoma Version 14.7.1_ on a mac laptop.
We had no problems on either machine. So I am not sure if it is worth waiting for an OS update.
Let me know if I can help or how we can proceed.
