### 2025-01-21 #6876 @britta-wstnr
@gouverneurp - mh, weird - I am on 14.6.1 and I cannot follow your procedure (I get stuck with not having the option to allow `run.command`. I will finally be able to update tonight and I will let you know if it works there. Then we'll see how to proceed given the outcome.
