### 2025-01-26 #6876 @gouverneurp
Hi @britta-wstnr,
Thanks again for the test and your help.

I tested the application on a current Mac Mini M4 with Sequoia and had no problems running the application:
![Image](https://github.com/user-attachments/assets/305aad21-f132-4ddf-8b50-2547d8313363)

As a reminder, you can find a detailed tutorial on how to launch the application here: [github link](https://github.com/gouverneurp/eVAS/blob/main/tutorials/run_on_mac.md)

Obviously, the process is a bit different depending on your Mac OS version, but it should work without any problems.
The option to "open anyway" in Sequoia can be found in this screenshot:
![Image](https://github.com/user-attachments/assets/8450af77-a905-4ea1-a6ff-28740a461f65)
[Image Source](https://media.idownloadblog.com/wp-content/uploads/2024/08/Apple-macOS-Sequoia-System-Settings-Security-Privacy-Section-Gatekeeper-Open-Anyway-button-annotation.jpg)

Honestly, I would like to refrain from stating that MacOs is not supported, as we could not reproduce any problems. I know that Apple's security restrictions hinder the usability of the application, but there is nothing we can do about it.

Please let me know how to proceed.

Best regards,
Philip
