### 2025-01-31 #6876 @britta-wstnr
Hi @gouverneurp,

I finally go the application to comply, thanks for the extra information.

Regarding the installation on MacOS, I have the following review feedback:
- [ ] please update the MacOS Download landing page https://gouverneurp.github.io/evas.html
 The page still says “You can dismiss the warning and press open”, which does not seem to be true on the two latest operating systems. Please also refer to your extra installation tutorial page here.
- [ ] Tutorial page https://github.com/gouverneurp/eVAS/blob/main/tutorials/mac_not_trusted.md
Please update this information for up-to-date MacOS systems (or add a second page). Judging based on the screenshots, this information is pre MacOS Ventura (when “Security & Privacy” got renamed to “Privacy & Security”), i.e. dating back to 2022 or earlier. As we have seen, the information does not work for newer MacOS versions.

Regarding the original review (see comment [here](https://github.com/openjournals/joss-reviews/issues/6876#issuecomment-2528469799 )):
Everything is indeed addressed by you, except:
- [ ] statement of need 	Is added to the README now. It might benefit from a few references.
Please consider at least adding references for the scales you mention.

Open points after trying eVAS:
- [x] From what I see, csv file headers are alright
- [x] setting up eVAS configuration is possible by downloading files. The config file is well documented which helps with tinkering with it.
- [ ] the application takes a long time for me to start up (over 30 seconds!). Please consider adding this to the README, so that users know that they have to be patient but that the application is not broken.
