### 2025-02-02 #6876 @gouverneurp
Hi @britta-wstnr,
Thanks again for the test and your help!

I have tried to follow your hints and requests. You can find my comments in the following:
- I have updated the MacOS Download landing page: [commit](https://github.com/gouverneurp/gouverneurp.github.io/commit/9a136091cbeac36838912383d201ce3bf68d7b17) [landing page](https://gouverneurp.github.io/evas.html)
- I have updated the tutorial page: [commit](https://github.com/gouverneurp/eVAS/commit/006d9090277e701d80b3901c4b2dd9253c37e962)
- More references have been added: [commit](https://github.com/gouverneurp/eVAS/commit/387908f7ac9cde77f7e35c2bd7524ae39437f23f)
- Long start up time: There is now information that the startup time can be long ([commit](https://github.com/gouverneurp/eVAS/commit/bfaf97fd52e35b8032813d7e14a9c0f0886b7a4a)). In addition, there is a new release with a leaner application (removed not needed packages like numpy) that could improve start up performance: [commit](https://github.com/gouverneurp/eVAS/commit/731ceaa2d602d74190aced7afba2bf3d92eb359f) [release 0.2.1](https://github.com/gouverneurp/eVAS/releases/tag/0.2.1)
