### 2025-02-03 #6876 @britta-wstnr
Thanks @gouverneurp. From my side, you have addressed all review points.

@samhforbes can you please have a look at this? Context: I stepped in for a reviewer who went non-responsive. My review is based on their first review (see [here](https://github.com/openjournals/joss-reviews/issues/6876#issuecomment-2235987208)) and carried out [here](https://github.com/openjournals/joss-reviews/issues/6876#issuecomment-2528469799) and [here](https://github.com/openjournals/joss-reviews/issues/6876#issuecomment-2627395017).
The other reviewer is content with the revisions after there review, see [here](https://github.com/openjournals/joss-reviews/issues/6876#issuecomment-2296641475).
Would you agree that we now can move onto the _Post review checklist_? I'd like to outsource this decision since I now doubled as a reviewer.
Thanks! 🙏
