### 2025-02-06 #6876 @samhforbes
Thanks @britta-wstnr I have gone through this carefully and am also happy that you've picked up the remains of the first review, and I'm also happy that these queries are answered. I think all good to push ahead to post review checklist. Thanks @haoxue-fan and @amitchell12 for taking the time to review.
