### 2025-02-07 #6876 @britta-wstnr
Great, thanks a lot, @samhforbes!
And indeed, many thanks to @haoxue-fan and @amitchell12 for your work and time, I and JOSS appreciate it! 🌻

@gouverneurp I will now double-check the paper for formatting and typos and such and get back to you. Once that is all in order, we'll start the pre-acceptance process.
