### 2025-02-12 #6876 @britta-wstnr
Hi @gouverneurp
I went over the paper again and have the following (formatting/style) comments:
- [ ] line 45: for the first citation, you do not follow the format of `lastname, firstname and lastname, firstname` in your `.bib` file which leads to the reference of "M. Szikszay".
- [ ] The [example paper](https://joss.readthedocs.io/en/latest/example_paper.html) stylizes softwares as `eVAS`, but since you are mainly reporting acronyms, it is okay in terms of parsing. I will leave that up to you.

Some sentences I stumbled over:
- line 15/16: I find "studies of e.g., pain, emotion, behaviour, anxiety and personality assessment" easier to parse than "studies of pain, emotion, behaviour, anxiety and personality assessment, for example."
- lines 20-22: I don't fully understand the last sentence. "This is particularly true..." What does "this" refer to? Grammatically it probably should be "open source software [...] is rare", but that does not make much sense.
- line 93: I find "If no error is logged, the empty log file will be removed" easier to parse.

 Lastly, I would like to mention that lines 52-54 of your paper can be perceived as a bit harsh. While factually true, it could be perceived as blaming the developers for "broken links". Personally, I would refrain from doing this since many reasons can lead to open source projects not supported any longer and I would rephrase the sentences. However, this is up to you as authors and you certainly do not have to follow my personal opinion - I just thought I'd mention it here.
