### 2025-02-16 #6876 @gouverneurp
Hi @britta-wstnr,

Thank you again for your feedback.
We have tried to address all of your comments:
- We fixed the citation: https://github.com/gouverneurp/eVAS/commit/8ed6003467a22cc2af970fd3d751a38147ae8e2b
- We stylized eVAS in the manuscript to keep the JOSS template: https://github.com/gouverneurp/eVAS/commit/3d1169f9f32228ef12cf203ba19f7d730bd68f0f
- We adjusted line 15: https://github.com/gouverneurp/eVAS/commit/4301bccc51fae89a5248aa0c6b0814f0181ab4bb
- We tried to make lines 20-22 easier to understand: https://github.com/gouverneurp/eVAS/commit/c886bae734ab3ffe7fdfa45b5489ece5b3174169
- We adjusted line 93: https://github.com/gouverneurp/eVAS/commit/89ddf489be491950aefe0056bc5232addbbed608
- Lines 52-54: We understand that this may seem harsh. We do not want to criticize any authors or open source software, as we know it is not easy to publish and maintain such projects. However, we feel that our statement sums up the current situation perfectly. So we would like to leave it as it is.

Let me know if there is anything else I can do to improve `eVAS`.
