### 2025-02-19 #6876 @britta-wstnr
Hi @gouverneurp looks good now, thanks!
Can you have a look at the **Author Tasks** for [this](https://github.com/openjournals/joss-reviews/issues/6876#issuecomment-2668182469) post-review checklist? Thanks a lot! 🙏
