### 2025-03-03 #6876 @britta-wstnr
Hi @gouverneurp can you have a look how you can add the version to figshare? Right now it says "Version 5", which I understand refers to the 5th upload on figshare, but it would be important to see which eVAS version this refers to. --- I am not very familiar with figshare, so I might be overlooking this.
To show you what I mean, take the right side of this page: https://zenodo.org/records/8322569. Each version gets its own DOI.

Thanks! 🙏
