### 2025-03-08 #6876 @gouverneurp
Hi @britta-wstnr,
To be honest, I have no idea how to set a version on figshare. So I added it to zenodo, you can find it here: https://zenodo.org/records/14992387. Version 0.2.1 is available at the following DOI: https://doi.org/10.5281/zenodo.14992387.
As soon as you mention that everything is OK, I will remove the figshare repository.
