### 2025-03-18 #6876 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
