### 2025-03-20 #6876 @gouverneurp
@britta-wstnr, who gives the command `editorialbot accept`? Do we need to provide this?
