### 2025-03-20 #6876 @danielskatz
This paper has encoding in the title. The authors should remove the encoding, and then the AEiC should reaccept it.
