### 2025-03-21 #6876 @gouverneurp
I removed the encoding in the title: https://github.com/gouverneurp/eVAS/commit/b49462c97517f10ec07e8fc313893ad9a4597ed4
