### 2025-03-21 #6876 @danielskatz
This looks good now - thanks again for the fix, @gouverneurp.

I'll leave this open for @Kevin-Mattheus-Moerman to check and officially close.
